fix graph bug
